+++ src/components/AI/IngredientList.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ListItem, UnorderedList } from "@chakra-ui/react";
-import { AIIngredient } from "../utils/types";
+import { AIIngredient } from "../../utils/types";
 
 type IngredientProps = { ingredient: AIIngredient };
