+++ src/components/AI/Loader.tsx
