+++ src/index.tsx
@@ -5,14 +5,14 @@ import { ChakraProvider } from "@chakra-ui/react";
 import reportWebVitals from "./reportWebVitals";
 import App from "./App";
 import Home from "./components/Home";
-import Register from "./components/Register";
-import Login from "./components/Login";
-import SearchAI from "./components/SearchAI";
-import SearchChoice from "./components/SearchChoice";
-import SavedRecipes from "./components/SavedRecipes";
+import Register from "./components/Login/Register";
+import Login from "./components/Login/Login";
+import SearchAI from "./components/AI/SearchAI";
+import SearchChoice from "./Search/SearchChoice";
+import SavedRecipes from "./components/SavedRecipe/SavedRecipes";
 import Layout from "./components/Layout";
-import SingleRecipePage from "./components/SingleRecipePage";
-import theme from "./CustomTheme.js";
+import SingleRecipePage from "./components/SingleResipePage/SingleRecipePage";
+import theme from "./CustomTheme";
 
 const root = ReactDOM.createRoot(document.getElementById("root")!);
 root.render(
